Adding Starbucks portal and portal scraper
This is not ready to merge, but I need help making the form typable. It's giving me a headache. :angry:

Their website's sign in portal is massive amounts of js only a maniac could go through. So I made a portal scraper also in this PR which rips their website, saves images local, and replaces those links with the new one. It also removes things like google analytics automatically.

Only problem I have is getting the damn form to allow input in the input boxes.

Feel free to fix or tell me to eff off. :wink: